% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acti_sleep_guiders.R
\name{acti_sleep_diary}
\alias{acti_sleep_diary}
\title{Identify a diary-defined sleep window}
\usage{
acti_sleep_diary(data, onset, wakeup)
}
\arguments{
\item{data}{A data frame containing timestamps.}

\item{onset, wakeup}{POSIXct values defining a diary sleep window.}
}
\value{
An \code{acti_sleep_guider} object. Missing diary times return an empty
window rather than imputing either boundary.
}
\description{
Identify a diary-defined sleep window
}
\examples{
\dontshow{if (rlang::is_installed("actimetrics")) withAutoprint(\{ # examplesIf}
data <- actimetrics::acti_count_data
diary_window <- acti_sleep_diary(
  data, onset = data$time[5], wakeup = data$time[30]
)
\dontshow{\}) # examplesIf}
}
