% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acti_sleep_guiders.R
\name{acti_sleep_guider_requirements}
\alias{acti_sleep_guider_requirements}
\title{Describe data requirements for sleep guiders}
\usage{
acti_sleep_guider_requirements()
}
\value{
A tibble describing the required data for each guider.
}
\description{
HDCZA and HorAngle take raw triaxial acceleration and create their own
regular epoch-level angle. All other guiders take a regularly spaced,
epoch-level data frame and do not require raw acceleration data.
}
\examples{
acti_sleep_guider_requirements()
}
