% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actiSleep.R
\name{actiSleep}
\alias{actiSleep}
\title{actiSleep algorithm}
\usage{
actiSleep(
  epoch_df,
  diary_df,
  stats_df,
  marker_df,
  time_start = NULL,
  time_stop = NULL,
  max_invalid_time = 4,
  add_na_tails = F,
  light_window = 120,
  push_window = 60
)
}
\arguments{
\item{epoch_df}{a data frame of epoch-level data}

\item{diary_df}{a data frame of diary data}

\item{stats_df}{a data frame of summary statistics}

\item{marker_df}{a data frame of button pushes}

\item{time_start}{a \code{POSIXct} object of the start time}

\item{time_stop}{a \code{POSIXct} object of the stop time}

\item{max_invalid_time}{maximum number of invalid hours per day; if over, the
data for a given day will be flagged as being "invalid."}
}
\value{
list of data.frames for marker start/stop times, epoch-level data,
epoch length, and invalid/removed day information.
}
\description{
This function implements the actiSleep algorithm. All of the data.frame
arguments are coming in similar format as output by \code{read_actigraphy()}.
}
