% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dayno.R
\name{add_dayno}
\alias{add_dayno}
\title{Create tracking day number}
\usage{
add_dayno(mydata, date_time_column, known_start)
}
\arguments{
\item{mydata}{a dataframe with a \code{POSIXct} column and an \code{ID} column}

\item{date_time_column}{character string corresponding to the datetime
column of \code{mydata}. Must be class \code{POSIXct}.}

\item{known_start}{start datetime of the observation period for which day
number is anchored. Must be class \code{POSIXct}.}
}
\value{
a data.frame with additional column \code{dayno}
}
\description{
This function creates a new column \code{dayno} that represents a day index
increment from noon-to-noon. A known start time must be supplied.
}
