% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SleepEstEachDay.R
\name{aggregate_dat}
\alias{aggregate_dat}
\title{Aggregate Accelerometer Data}
\usage{
aggregate_dat(datain, cutnum = 1, unit = c("min", "sec"), tz = "GMT")
}
\arguments{
\item{datain}{raw ActiGraph GT3X data}

\item{cutnum}{quantity of unit to serve as interval, default is 1}

\item{unit}{time unit to use}

\item{tz}{timezone, default is GMT}
}
\value{
data frame of accelerometer data aggregated by intervals
}
\description{
Obtain table from raw accelerometer table in 1-min intervals
}
