% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SleepEstEachDay.R
\name{cal_mean_idx}
\alias{cal_mean_idx}
\title{Calculate Mean Index}
\usage{
cal_mean_idx(i, datain, idx_start, idx_end)
}
\arguments{
\item{i}{segment index}

\item{datain}{activity data, must be a data frame}

\item{idx_start}{array of segment start indices}

\item{idx_end}{array of segment end indices}
}
\value{
mean value of segment indices
}
\description{
Calculate the mean index of a segment
}
