% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stats.R
\name{calculate_stats}
\alias{calculate_stats}
\title{Calculate statistics}
\usage{
calculate_stats(all_markers, epochs_f, ep_factor)
}
\arguments{
\item{all_markers}{description}

\item{epochs_f}{description}

\item{ep_factor}{description}
}
\description{
This function calculates statistics from marker and epoch data
}
