% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegmentorInstall.r
\name{.onAttach}
\alias{.onAttach}
\title{Announce Segmentor3IsBack has Loaded}
\usage{
.onAttach(lib, pkg)
}
\arguments{
\item{lib}{FILLER TEXT}

\item{pkg}{FILLER TEXT}
}
\description{
Print out a statement to the console when Segmentor3IsBack is
successfully loaded
}
\details{
This code was written by Alice Cleynen, Guillem Rigaill, and Michel Koskas
as part of the Segmentor3IsBack package, which is no longer in CRAN. It has
been imported into the ActiSleep package to ensure this package's longevity.
}
