% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_main.R
\name{find_main}
\alias{find_main}
\title{Find main/nap}
\usage{
find_main(mydata, col1, col2)
}
\arguments{
\item{mydata}{epoch level data.frame}

\item{col1}{start column name string}

\item{col2}{stop column name string}
}
\description{
find main/nap
}
