% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearby_rest_functions.R
\name{find_nearby_rest}
\alias{find_nearby_rest}
\title{Find naps from rest actigraphy data}
\usage{
find_nearby_rest(mydata, nmin = 30)
}
\arguments{
\item{mydata}{a data.frame of start/end rest that has a \code{dayno} column}

\item{nmin}{number of minutes in-between intervals}
}
\value{
a data.frame with added \code{interval_match} variable denoting the row
index of the rest interval within \code{nmin} minutes of the current row.
}
\description{
This function creates a new variable called \code{interval_match} which indicates
the row index of rest event should be merged with another. Rest intervals
get merged if they're within a certain number of minutes together.
}
