% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearby_rest_functions.R
\name{find_nocturnal_awakening}
\alias{find_nocturnal_awakening}
\title{Find nocturnal from rest actigraphy data}
\usage{
find_nocturnal_awakening(mydata)
}
\arguments{
\item{mydata}{a data.frame of start/end rest that has a \code{dayno} column}
}
\value{
a data.frame with added \code{interval_match} variable denoting the row
index of the rest interval within \code{nmin} minutes of the current row.
}
\description{
A function that identifies rest intervals that are separated by a nocturnal
awakening. Rather than defining nocturnal statically, this function uses the
median start/end time of the longest rest intervals to determine night time.
}
