% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_afternoon_rest.R
\name{fix_afternoon_rest}
\alias{fix_afternoon_rest}
\title{Fix afternoon rest}
\usage{
fix_afternoon_rest(df, updated_rest, end_nm)
}
\arguments{
\item{df}{data.frame with epoch-level data}

\item{updated_rest}{data.frame with start and end times of rest intervals.
If some of these pass noon, the day number will be adjusted.}

\item{end_nm}{character string with the name of the end time column in
\code{updated_rest}}
}
\value{
a data.frame of epoch-level data with updated tracking day numbers.
If rest passes noon, then the day number are adjusted to the day in which the
rest began.
}
\description{
This function adjusts the day number of rest data that passes noon.
}
