% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_actigraphy_diary_marker.R
\name{join_actigraphy_diary_marker}
\alias{join_actigraphy_diary_marker}
\title{Joins cleaned actigraphy, sleep diary, and marker rest intervals}
\usage{
join_actigraphy_diary_marker(updated_sleep_df, cleaned_marker_df)
}
\arguments{
\item{updated_sleep_df}{dataframe that is outputted from \code{update_sleep()}--
this dataset should have already had engines/cabooses taken care of}

\item{cleaned_marker_df}{dataframe that is outputted from from \code{find_markers}
-- this dataset should already have been run through \code{make_dayno_jg}}
}
\value{
a dataframe with columns: ID, dayno, Type (Main vs. Nap),
actigraphy.Start, diary.Start, marker.Start, actigraphy.Stop, diary.Stop,
marker.Stop
}
\description{
This function takes the cleaned actigraphy rest dataset returned from
\code{update_sleep()}, cleaned sleep diary data from \code{sleep_diary_formatting},
cleaned marker data from \code{find_markers} and joins the datasets together
}
