% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_actogram.R
\name{plot_actogram}
\alias{plot_actogram}
\title{Plot actogram}
\usage{
plot_actogram(
  sd_data,
  all_markers,
  dmin,
  dmax,
  filter_naps = F,
  db = "60 min",
  flags,
  epoch_length = 1
)
}
\arguments{
\item{sd_data}{epoch level data.frame}

\item{all_markers}{marker data.frame}

\item{dmin}{minimum day number}

\item{dmax}{maximum day number}

\item{filter_naps}{true or false on filtering naps}

\item{db}{default \code{'60 min'} tick amount on x-axis}

\item{flags}{flag data.frame}

\item{epoch_length}{epoch length in minutes}
}
\value{
a list of data.frames corresponding to properties, statistics, markers,
epochs-level.
}
\description{
use after actiSleep
}
