% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_actigraphy.R
\name{read_actigraphy}
\alias{read_actigraphy}
\title{Read actigraphy data}
\usage{
read_actigraphy(file_name)
}
\arguments{
\item{file_name}{character string of the file name. It should be a .csv file.}
}
\value{
a list of data.frames corresponding to properties, statistics, markers,
epochs-level.
}
\description{
This function takes in an actigraphy file and returns a list of data.frames
that include the properties, summary statistics, button presses (marker),
and epoch-level data.
}
