% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agd.R
\name{read_agd}
\alias{read_agd}
\title{Read in AGD filedata}
\usage{
read_agd(file, tz = "GMT", sec = 10)
}
\arguments{
\item{file}{AGD data file}

\item{tz}{timezone, default is GMT}

\item{sec}{time interval used to set date}
}
\value{
list of AGD filenames and data
}
\description{
Obtain a list of AGD filenames and data
}
