% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SleepEstEachDay.R
\name{SearchSleepSeg}
\alias{SearchSleepSeg}
\title{Find specific sleep segment}
\usage{
SearchSleepSeg(
  date,
  Y,
  idx_start,
  idx_end,
  NoPA_cut = 0.7,
  sleep_mins = 30,
  SI_i,
  tz = "GMT"
)
}
\arguments{
\item{date}{date of interest}

\item{Y}{vector containing validated sleep counts}

\item{idx_start}{start index of sleep segment}

\item{idx_end}{end index of sleep segment}

\item{NoPA_cut}{percent of 0 counts to define NoPA segments}

\item{sleep_mins}{threshold number of minutes to define sleep segment}

\item{SI_i}{estimated sleep intervals based on sleep diary data}

\item{tz}{timezone, default is GMT}
}
\value{
list of validated sleep segments
}
\description{
Obtain specific sleep segment based on day, indices, etc.
}
