% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegmentorClasses.R
\docType{class}
\name{Segmentor-class}
\alias{Segmentor-class}
\title{This code was written by Alice Cleynen, Guillem Rigaill, and Michel Koskas
as part of the Segmentor3IsBack package, which is no longer in CRAN. It has
been imported into the ActiSleep package to ensure this package's longevity.}
\description{
This code was written by Alice Cleynen, Guillem Rigaill, and Michel Koskas
as part of the Segmentor3IsBack package, which is no longer in CRAN. It has
been imported into the ActiSleep package to ensure this package's longevity.
}
