% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentor3IsBack.R
\name{Segmentor}
\alias{Segmentor}
\title{PDP-based Segmenting Algorithm}
\usage{
Segmentor(
  data = numeric(),
  model = 1,
  Kmax = 15,
  phi = numeric(),
  m = numeric(),
  keep = FALSE,
  bounds = c(0, 0),
  compress = TRUE
)
}
\arguments{
\item{data}{FILLER TEXT}

\item{model}{number 1-5 indicating which distribution model to use}

\item{Kmax}{FILLER TEXT}

\item{phi}{FILLER TEXT}

\item{m}{FILLER TEXT}

\item{keep}{FILLER TEXT}

\item{bounds}{FILLER TEXT}

\item{compress}{FILLER TEXT}
}
\description{
This code was written by Alice Cleynen, Guillem Rigaill, and Michel Koskas
as part of the Segmentor3IsBack package, which is no longer in CRAN. It has
been imported into the ActiSleep package to ensure this package's longevity.
}
