% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SleepDiary1Day.R
\docType{data}
\name{SleepDiary1Day}
\alias{SleepDiary1Day}
\title{Sleep Diary 1-Day Data}
\format{
a data frame with 1 row and 2 variables
}
\usage{
data("SleepDiary1Day")
}
\description{
A small dataset containing sleep diary data for a single day
and single subject

\itemize{
  \item bed. time at which the subject reported going to sleep on day 1
  \item wake. time at which the subject reported waking up on day 1
}
}
\keyword{datasets}
