% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearby_rest_functions.R
\name{update_sleep}
\alias{update_sleep}
\title{Update rest interval}
\usage{
update_sleep(mydata)
}
\arguments{
\item{mydata}{a data.frame of start/end rest that has a \code{dayno} column}
}
\value{
a data.frame with updated start and end times of rest intervals
}
\description{
This function updates the start and end times of rest intervals as determined
by \code{find_nearby_rest()}.
}
