% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccelData.R
\docType{data}
\name{AccelData}
\alias{AccelData}
\title{Accelerometry Data}
\format{
a data frame with 1351 rows and 11 variables
}
\usage{
data("AccelData")
}
\description{
A dataset containing accelerometry data for a single subject in
1-minute epochs, includes vector magnitude and other measurements

\itemize{
  \item date. time and date of single measurement
  \item axis1. x-axis movement
  \item axis2. y-axis movement
  \item axis3. z-axis movement
  \item steps. steps taken by subject
  \item lux. amount of light present, measured in lux
  \item inclineOff.
  \item inclineStanding.
  \item inclineSitting.
  \item inclineLying.
  \item VM. vector magnitude, calculated from x-, y-, z-axis measurements
}
}
\keyword{datasets}
