% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acti_sleep_guiders.R
\name{acti_sleep_setwindow}
\alias{acti_sleep_setwindow}
\title{Identify a fixed daily sleep window}
\usage{
acti_sleep_setwindow(data, start_hour = 22, end_hour = 8)
}
\arguments{
\item{data}{A data frame containing timestamps.}

\item{start_hour, end_hour}{Start and end clock hours (0--24).}
}
\value{
An \code{acti_sleep_guider} object.
}
\description{
Identify a fixed daily sleep window
}
\examples{
data <- actimetrics::acti_count_data
fixed_window <- acti_sleep_setwindow(data, start_hour = 22, end_hour = 8)
}
